Display work order details when clicking on a work order on the service request management page

+++ app/Http/Livewire/EmployeeServiceRequestManage.php
@@ -4,12 +4,14 @@ namespace App\Http\Livewire;
 
 use App\Models\ServiceRequest;
 use App\Models\Tenant;
+use App\Models\WorkDetails;
 use App\Models\WorkOrder;
 use Livewire\Component;
 
 class EmployeeServiceRequestManage extends Component
 {
     public $request;
+    public $currentWorkOrderId;
 
     public function toggleComplete()
     {
@@ -18,6 +20,11 @@ class EmployeeServiceRequestManage extends Component
         } else {
             $this->request->update(['completed_date' => now()]);
         }
+    }
+
+    public function displayWorkOrderDetails($workOrderID)
+    {
+        $this->currentWorkOrderId = $workOrderID;
     }
 
     public function newWorkOrder()
@@ -30,8 +37,10 @@ class EmployeeServiceRequestManage extends Component
     public function deleteWorkOrder($id)
     {
         $workOrder = WorkOrder::find($id);
-
-        $workOrder->delete();
+        
+        if (! $workOrder->hasWorkDetails()) {
+            $workOrder->delete();
+        }
     }
 
     public function render()
@@ -39,6 +48,8 @@ class EmployeeServiceRequestManage extends Component
         return view('livewire.employee-service-request-manage', [
 
             'workOrders' => WorkOrder::where('service_request_id', $this->request->id)->get(),
+
+            'workOrderDetails' => WorkDetails::where('work_order_id', $this->currentWorkOrderId)->get(),
 
         ]);
     }
